+++ trade-finance-manager-ui/server/controllers/case/index.js
@@ -205,108 +205,6 @@ const getCaseDocuments = async (req, res) => {
   });
 };
 
-
-const getCaseParties = async (req, res) => {
-  const dealId = req.params._id; // eslint-disable-line no-underscore-dangle
-  const deal = await api.getDeal(dealId);
-
-  if (!deal) {
-    return res.redirect('/not-found');
-  }
-
-  return res.render('case/parties/parties.njk', {
-    deal: deal.dealSnapshot,
-    tfm: deal.tfm,
-    activePrimaryNavigation: 'manage work',
-    activeSubNavigation: 'parties',
-    dealId,
-    user: req.session.user,
-  });
-};
-
-const getPartyDetails = (partyType) => (
-  async (req, res) => {
-    const dealId = req.params._id; // eslint-disable-line no-underscore-dangle
-    const deal = await api.getDeal(dealId);
-
-    if (!deal) {
-      return res.redirect('/not-found');
-    }
-
-    return res.render(`case/parties/edit/${partyType}-edit.njk`, {
-      activePrimaryNavigation: 'manage work',
-      activeSubNavigation: 'parties',
-      deal: deal.dealSnapshot,
-      tfm: deal.tfm,
-      dealId,
-      user: req.session.user,
-    });
-  }
-);
-
-const getExporterPartyDetails = getPartyDetails('exporter');
-const getBuyerPartyDetails = getPartyDetails('buyer');
-const getAgentPartyDetails = getPartyDetails('agent');
-const getIndemnifierPartyDetails = getPartyDetails('indemnifier');
-
-const getBondIssuerPartyDetails = async (req, res) => {
-  const dealId = req.params._id; // eslint-disable-line no-underscore-dangle
-  const deal = await api.getDeal(dealId);
-
-  if (!deal) {
-    return res.redirect('/not-found');
-  }
-
-  return res.render('case/parties/edit/bonds-issuer-edit.njk', {
-    activePrimaryNavigation: 'manage work',
-    activeSubNavigation: 'parties',
-    deal: deal.dealSnapshot,
-    user: req.session.user,
-  });
-};
-
-
-const getBondBeneficiaryPartyDetails = async (req, res) => {
-  const dealId = req.params._id; // eslint-disable-line no-underscore-dangle
-  const deal = await api.getDeal(dealId);
-
-  if (!deal) {
-    return res.redirect('/not-found');
-  }
-
-  return res.render('case/parties/edit/bonds-beneficiary-edit.njk', {
-    activePrimaryNavigation: 'manage work',
-    activeSubNavigation: 'parties',
-    deal: deal.dealSnapshot,
-    user: req.session.user,
-  });
-};
-
-const postPartyDetails = (partyType) => (
-  async (req, res) => {
-    const dealId = req.params._id; // eslint-disable-line no-underscore-dangle
-
-    const deal = await api.getDeal(dealId);
-
-    if (!deal) {
-      return res.redirect('/not-found');
-    }
-
-    const update = {
-      [partyType]: req.body,
-    };
-
-    await api.updateParty(dealId, update);
-
-    return res.redirect(`/case/${dealId}/parties`);
-  }
-);
-
-const postExporterPartyDetails = postPartyDetails('exporter');
-const postBuyerPartyDetails = postPartyDetails('buyer');
-const postAgentPartyDetails = postPartyDetails('agent');
-const postIndemnifierPartyDetails = postPartyDetails('indemnifier');
-
 const postTfmFacility = async (req, res) => {
   const { facilityId, ...facilityUpdateFields } = req.body;
   const dealId = req.params._id; // eslint-disable-line no-underscore-dangle
@@ -341,16 +239,5 @@ export default {
   putCaseTask,
   getCaseFacility,
   getCaseDocuments,
-  getCaseParties,
-  getExporterPartyDetails,
-  getBuyerPartyDetails,
-  getAgentPartyDetails,
-  getIndemnifierPartyDetails,
-  getBondIssuerPartyDetails,
-  getBondBeneficiaryPartyDetails,
-  postExporterPartyDetails,
-  postBuyerPartyDetails,
-  postAgentPartyDetails,
-  postIndemnifierPartyDetails,
   postTfmFacility,
 };
