+++ trade-finance-manager-ui/server/routes/case/index.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import caseController from '../../controllers/case';
+import partiesController from '../../controllers/case/parties';
 import underwritingController from '../../controllers/case/underwriting';
 import activityController from '../../controllers/case/activity';
 
@@ -15,25 +16,25 @@ router.post('/:_id/tasks/:groupId/:taskId', caseController.putCaseTask);
 
 router.get('/:_id/facility/:facilityId', caseController.getCaseFacility);
 
-router.get('/:_id/parties', caseController.getCaseParties);
+router.get('/:_id/parties', partiesController.getCaseParties);
 
-router.get('/:_id/parties/exporter', caseController.getExporterPartyDetails);
-router.post('/:_id/parties/exporter', caseController.postExporterPartyDetails);
+router.get('/:_id/parties/exporter', partiesController.getExporterPartyDetails);
+router.post('/:_id/parties/exporter', partiesController.postExporterPartyDetails);
 
-router.get('/:_id/parties/buyer', caseController.getBuyerPartyDetails);
-router.post('/:_id/parties/buyer', caseController.postBuyerPartyDetails);
+router.get('/:_id/parties/buyer', partiesController.getBuyerPartyDetails);
+router.post('/:_id/parties/buyer', partiesController.postBuyerPartyDetails);
 
-router.get('/:_id/parties/agent', caseController.getAgentPartyDetails);
-router.post('/:_id/parties/agent', caseController.postAgentPartyDetails);
+router.get('/:_id/parties/agent', partiesController.getAgentPartyDetails);
+router.post('/:_id/parties/agent', partiesController.postAgentPartyDetails);
 
-router.get('/:_id/parties/indemnifier', caseController.getIndemnifierPartyDetails);
-router.post('/:_id/parties/indemnifier', caseController.postIndemnifierPartyDetails);
+router.get('/:_id/parties/indemnifier', partiesController.getIndemnifierPartyDetails);
+router.post('/:_id/parties/indemnifier', partiesController.postIndemnifierPartyDetails);
 
-router.get('/:_id/parties/bond-issuer', caseController.getBondIssuerPartyDetails);
-router.post('/:_id/parties/bond-issuer', caseController.postTfmFacility);
+router.get('/:_id/parties/bond-issuer', partiesController.getBondIssuerPartyDetails);
+router.post('/:_id/parties/bond-issuer', partiesController.postTfmFacility);
 
-router.get('/:_id/parties/bond-beneficiary', caseController.getBondBeneficiaryPartyDetails);
-router.post('/:_id/parties/bond-beneficiary', caseController.postTfmFacility);
+router.get('/:_id/parties/bond-beneficiary', partiesController.getBondBeneficiaryPartyDetails);
+router.post('/:_id/parties/bond-beneficiary', partiesController.postTfmFacility);
 
 router.get('/:_id/activity', activityController.getActivity);
